requirements.txt
beautifulsoup4
requests
openai
PySocks